backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
PyMySQL==1.1.0
mysqlclient==2.2.0
requests==2.31.0
psutil==5.9.5
python-dotenv==1.0.0
gunicorn==21.2.0
cryptography==41.0.7
Werkzeug==2.3.7
